+++ tests/Queue/e2e/Adapter/WorkermanTest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Tests\E2E\Adapter;
+
+use Utopia\Queue\Client;
+use Utopia\Queue\Connection\Redis;
+
+class WorkermanTest extends Base
+{
+    protected function getClient(): Client
+    {
+        $connection = new Redis('redis', 6379);
+        $client = new Client('workerman', $connection);
+
+        return $client;
+    }
+}
